Create github page to make cards.pdf available
This is done via a github workfloaw, that is run after the
LaTeX build workflow is done, when that happened on the main branch.

This has the drawback, that it currently doesn't show as check on the commit.

With this commit merged, we have a small website at https://toxaris.github.com/pathfinder2-cards
and our cards.pdf is available at https://toxaris.github.io/pathfinder2-cards/cards.pdf.
I tested this on a fork, which appears to work (see https://fkz.github.io/pathfinder2-cards).

This allows to view the generated cards.pdf directly on a website.